Rename mock service and update imports
