+++ app/services/index.ts
@@ -1,4 +1,4 @@
 export * as api from './api';
-export * as mock from './mockDatabaseService';
+export * as storage from './storageService';
 export * as db from './databaseService';
 export { addNode, removeNode, stopNode, startNode } from './api';
